Move add_library_item_topics
