+++ library/helpers.py
@@ -1,5 +1,8 @@
 import typing
 
+from facets.models import Topic
+from library.models import LibraryItemTopic
+
 
 if typing.TYPE_CHECKING:
     from library.models import LibraryItem
@@ -38,3 +41,29 @@ def create_querystring_from_facets(
     # Join facets into a querystring
     # placing an ampersand between each key/value pair
     return "&".join(f"{key}={value}" for key, value in facets.items())
+
+def add_library_item_topics(
+    library_item: "LibraryItem",
+    topics: str,
+) -> None:
+    """Add topics to a library item."""
+
+    topics = topics.split(";")
+
+    for topic_title in topics:
+        if topic_title == "":
+            continue
+
+        try:
+            topic = Topic.objects.get(
+                title=topic_title,
+            )
+        except Topic.DoesNotExist:
+            print(f"Topic '{topic_title}' does not exist")
+
+        library_item_topic = LibraryItemTopic(
+            library_item=library_item,
+            topic=topic,
+        )
+
+        library_item_topic.save()
